Revert "Add `ROG_MAXIMUS_XII_Z490_FORMULA` support"
This reverts commit 08da06969bf0c42c190d8aeaab930464748c468a.

+++ LibreHardwareMonitorLib/Interop/Ftd2xx.cs
@@ -161,7 +161,7 @@ internal static class Ftd2xx
     [StructLayout(LayoutKind.Sequential)]
     internal struct FT_HANDLE
     {
-        private readonly IntPtr _handle;
+        private readonly uint _handle;
     }
 
     [StructLayout(LayoutKind.Sequential)]
